requirements.txt
pandas==1.1.5
numpy==1.17.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
Shapely==1.7.1
pycocotools==2.0.2
imgaug==0.4.0
p-tqdm==1.3.3
fonttools==4.27.1
torchvision==0.4.1a0+d94043a
Pillow==8.2.0
matplotlib==3.3.4
albumentations==1.1.0
scipy==1.5.4
